+++ src/LinearSolvers/LinearSolvers.jl
@@ -6,13 +6,11 @@ using LinearAlgebra
 using SparseArrays
 using SparseMatricesCSR
 using BlockArrays
-using IterativeSolvers
 using FillArrays
 
 using Gridap
 using Gridap.Helpers, Gridap.Algebra, Gridap.CellData, Gridap.Arrays, Gridap.FESpaces, Gridap.MultiField
 using PartitionedArrays
-using GridapPETSc
 
 using GridapDistributed
 using GridapSolvers.MultilevelTools
@@ -50,11 +48,6 @@ include("Krylov/GMRESSolvers.jl")
 include("Krylov/FGMRESSolvers.jl")
 include("Krylov/MINRESSolvers.jl")
 
-include("PETSc/PETScUtils.jl")
-include("PETSc/PETScCaches.jl")
-include("PETSc/ElasticitySolvers.jl")
-include("PETSc/HipmairXuSolvers.jl")
-
 include("IdentityLinearSolvers.jl")
 
 include("LinearSolverFromSmoothers.jl")
@@ -64,7 +57,6 @@ include("SymGaussSeidelSmoothers.jl")
 include("RichardsonLinearSolvers.jl")
 
 include("GMGLinearSolvers.jl")
-include("IterativeLinearSolvers.jl")
 include("SchurComplementSolvers.jl")
 include("MatrixSolvers.jl")
 include("SchwarzLinearSolvers.jl")