+++ test/Applications/NavierStokesGMG.jl
@@ -34,7 +34,6 @@ using Gridap.ReferenceFEs, Gridap.Algebra, Gridap.Geometry, Gridap.FESpaces
 using Gridap.CellData, Gridap.MultiField, Gridap.Algebra
 using PartitionedArrays
 using GridapDistributed
-using GridapP4est
 
 using GridapSolvers
 using GridapSolvers.LinearSolvers, GridapSolvers.MultilevelTools
@@ -135,7 +134,7 @@ function main(distribute,np,nc,np_per_level)
     tests_u,jac_u,graddiv,qdegree;is_nonlinear=true
   )
   restrictions = setup_patch_restriction_operators(
-    tests_u,prolongations,graddiv,qdegree;solver=IS_ConjugateGradientSolver(;reltol=1.e-6)
+    tests_u,prolongations,graddiv,qdegree;solver=CGSolver(JacobiLinearSolver())
   )
   gmg = GMGLinearSolver(
     trials_u,tests_u,biforms,