+++ test/LinearSolvers/GMGTests.jl
@@ -3,14 +3,12 @@ module GMGTests
 using MPI
 using Test
 using LinearAlgebra
-using IterativeSolvers
 using FillArrays
 
 using Gridap
 using Gridap.ReferenceFEs, Gridap.Algebra
 using PartitionedArrays
 using GridapDistributed
-using GridapP4est
 
 using GridapSolvers
 using GridapSolvers.LinearSolvers
@@ -50,7 +48,7 @@ function gmg_driver_from_mats(t,parts,mh,spaces,qdegree,smoothers,biform,liform,
   tests, trials = spaces
 
   restrictions, prolongations = setup_transfer_operators(
-    trials, qdegree; mode=:residual, solver=IS_ConjugateGradientSolver(;reltol=1.e-6)
+    trials, qdegree; mode=:residual, solver=CGSolver(JacobiLinearSolver())
   )
 
   smatrices, A, b = compute_hierarchy_matrices(trials,tests,biform,liform,qdegree)
@@ -91,7 +89,7 @@ function gmg_driver_from_weakform(t,parts,mh,spaces,qdegree,smoothers,biform,lif
   tests, trials = spaces
 
   restrictions, prolongations = setup_transfer_operators(
-    trials, qdegree; mode=:residual, solver=IS_ConjugateGradientSolver(;reltol=1.e-6)
+    trials, qdegree; mode=:residual, solver=CGSolver(JacobiLinearSolver())
   )
 
   A, b = with_level(mh,1) do _